pyproject.toml
[tool.poetry]
name = "inteligencja-proj2"
version = "0.1.0"
description = ""
authors = ["Mateusz Kurowski <m.kurowski.023@studms.ug.edu.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
twscrape = "^0.12"
numpy = "^1.26.4"
nltk = "^3.8.1"
autocorrect = "^2.6.1"
scikit-learn = "^1.5.0"
text2emotion = "^0.0.5"
emoji = "1.7.0"
transformers = "^4.41.2"
tf-keras = "^2.16.0"
pillow = "^10.3.0"
wordcloud = "^1.9.3"
pandas-stubs = "^2.2.2.240603"
stopwordsiso = "^0.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
